backend/requirements.txt
fastapi
uvicorn
python-uinput
websockets
pytest
httpx
